profile page completed

+++ backend/controllers/deleteUser.js
@@ -2,7 +2,16 @@ const User = require('../model/UserSchema');
 
 const deleteUser = async(req,res) => {
     try{
-        const email = req.body.username;
+        const {username} = req.body;
+        const email = username;
+
+        if(!email){
+            return res.status(500).json({
+                message: "Something went wrong (email not recieved)",
+                success: false,
+            })
+        }
+
         console.log("Username of deleting user : ",req.body);
         const deleteUser = await User.deleteOne({username: email});
         if(deleteUser){
